+++ utils/src/main/scala/com.olegych.scastie/SbtTask.scala
@@ -4,7 +4,7 @@ import api._
 
 import akka.actor.ActorRef
 
-case class SbtTask(id: Long,
+case class SbtTask(id: Int,
                    inputs: Inputs,
                    ip: String,
                    progressActor: ActorRef)
